pwiz-sharp/BiblioSpec: round of port-fidelity fixes (cpp parity)
A cluster of pwiz-sharp BiblioSpec / MsData / vendor-reader bugs found
exercising the port end-to-end against the BiblioSpec TestData corpus
and Skyline's TestTutorial fixtures. Each fix anchors on the cpp call
site it mirrors.

Lazy mzXML reader
- pwiz/src/MsData/MzxmlIndexFooter.cs: validate the trailer-supplied
  index by probing start / middle / end offsets — each must land on a
  <scan num="N"> tag whose num matches the index entry. Old TPP / ReAdW
  indexes routinely lie (mid-base64-peak offsets, off-by-many-KiB
  drift), and the C# lazy reader silently surfaced the wrong scan
  (often an MS1 in place of an MS2). cpp's HandlerIndex prints
  "<index> not found." and rebuilds via a forward sweep; the C# port
  now falls back to the in-line index build when validation fails.
- pwiz/src/MsData/Mzxml/SpectrumList_Mzxml.cs: SeekToScanTag byte-scans
  for "<scan " before XmlReader.ReadToFollowing. Even well-formed
  indexes record offsets a few bytes before the scan they describe
  (start of preceding whitespace, inside a trailing "/>" /
  "</dataProcessing>"); .NET XmlReader chokes on unmatched close tags
  where cpp's SAX parser walks past them.

PepXMLreader
- WrapStreamWithMissingXsiDecl splices
  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" into the root
  element of old TPP interact-* / SEQUEST .pepXML files that use
  xsi:type / xsi:schemaLocation WITHOUT declaring the prefix. cpp's
  expat tolerates it; .NET XmlReader rejects 'xsi' as undeclared.
  Detection anchors on xmlns:xsi= (not bare xmlns:xsi, which would
  false-positive on xmlns:xsibad=) with a \s*= regex fallback; 8 KiB
  head read loops on short reads.
- MapToFileCoordinates + CompositeReadStream.InjectionLine/Column/
  Length translate XmlException line/position back to on-disk file
  coordinates so an editor jumping to the reported position lands at
  the real problem, not on injected whitespace. CompositeReadStream
  doesn't own its tail FileStream, so Parse disposes both explicitly
  in finally.

BlibBuilder + Verbosity
- Score-lookup mode (-s -t) emits per-file
  <filename>\n<scoreType>\t<probType>\n blocks per cpp BlibBuild.cpp:
  147-155 + 215-221. EmitScoreLookupFailureRow centralizes the UNKNOWN
  fallback so every filename header is followed by at least one
  scoreType row — even when the reader's ctor throws, no factory
  accepts the extension, or the file is unreadable. Per-file catch
  arms no longer set success=false in lookup mode (cpp BlibBuild.cpp:
  253 guards the same way), so callers gating on exit code don't drop
  a mixed batch over one bad file.
- Bare .group (not .group.xml) short-circuits before the reader-factory
  loop: normal mode emits cpp's "must be converted to .group.xml"
  guidance instead of XmlException; lookup mode emits
  ProteinPilotReader.GetScoreTypesHelper() without instantiating a
  reader (cpp BlibBuild.cpp:176-184).
- Verbosity.WriteErrorLines (cpp BlibBuild.cpp:78) splits multi-line
  BlibException messages and prefixes EVERY line with "ERROR: " —
  line-buffer-filtered downstream consumers otherwise keep only the
  first prefixed line and lose the trailing "Run with the -E flag"
  anchor that BiblioSpecLite-shape error parsers rely on.
  BlibBuild/Program.cs routes all per-file errors through it.

BlibSearch
- Decoy generation ported against cpp SearchLibrary.cpp:251-261 +
  270-284: GenerateDecoySpectra creates DecoysPerTarget copies per
  target per CircShift / NewDecoy + ShiftRawSpectrum.
- Peak-processing loop iterates from a startDecoyIdx captured BEFORE
  GenerateDecoySpectra, not the target-list startIdx — handles
  NewDecoy returning null (rejected as too few peaks) and multi-
  library inputs where decoys from earlier libraries are already in
  _cachedDecoySpectra.
- Accept-and-ignore dev flags (--compute-p-values etc.) no longer
  eat the next argv element when invoked without an inline value.
  Inline only.

DotProduct + PeakProcessor
- DotProduct.cs: drop the explicit (double) cast on the matched-mz
  branch. cpp DotProduct.cpp:68-70 uses float*float on matched peaks
  (pow returns double on unmatched/drain) — the asymmetry is load-
  bearing for tie-breaks in BlibFilter's representative selection.
  Filter's TASEFDSAIAQDK group: 0.7524627... ties between ids 50 and
  116, cpp picks 50; the cast inflated precision enough to flip the
  pick.
- PeakProcessor.cs: UpperBoundByMz mirrors cpp compPeakMz, which
  treats (mz, 0) as less than (mz, positive_intensity) — so
  std::upper_bound stops on the first peak at search_mz with
  intensity > 0. Strict mz > search_mz walked one position further;
  combined with RemovePrecursorPeaks's "last++" expansion that erased
  one extra binned peak past the +5 boundary (Search_Binning).

BlibUtils
- GetExeDirectory prefers AppContext.BaseDirectory; falls back
  through Process.MainModule.FileName then
  Assembly.GetEntryAssembly().Location. Assembly.Location is empty
  under PublishSingleFile=true, so modifications.xml / unimod.xml
  lookups broke for single-file consumers.

Vendor csproj
- Sciex / Shimadzu: vendor DLLs get CopyToPublishDirectory=Never —
  ProjectReference already propagates them via build output, so the
  Content path would otherwise re-add them and trigger NETSDK1152
  ("BundleRelativePath ... has duplicates") on
  -p:PublishSingleFile=true. Sciex .config files split into a separate
  Content block that DOES publish: the three Unity DI configs aren't
  propagated by ProjectReference (only compiled outputs are), and the
  SCIEX SDK reads them from the working directory — without them
  request handlers NPE inside the container.

Test suite hookups
- FilterTests.Filter: flipped from Inconclusive to live; passes
  against the cpp golden.
- SearchTests.Search_Decoy + Search_Binning: flipped live; goldens
  match cpp output.
- BuildTests.Ssl + Ssl_Crosslink + Omssa: flipped live (the runner's
  Unicode-rename harness mirrors cpp's --unicode pre-step).
- BlibBuild.csproj: ship modifications.xml + unimod.xml as Content
  next to the exe (cpp parity — MaxQuantReader.cpp:177 and
  mzTabReader.cpp:66 both fall back to <exeDir>/<file>).

Suite: 128 passed / 0 failed / 6 skipped (was 122 / 0 / 12). The 6
remaining Inconclusives are the Mascot cluster — port lands in a
follow-up commit with the msparser shim.

+++ pwiz-sharp/Tools/BiblioSpec/src/BiblioSpec/DotProduct.cs
@@ -29,8 +29,12 @@ public static class DotProduct
     /// // angle = expRefIntSum / sqrt(expIntSqSum * refIntSqSum)
     /// // if angle is NaN (both sums zero), angle = 0.
     /// </code></para>
-    /// <para>cpp uses <c>double</c> for the multiplication explicitly (cpp DotProduct.cpp:93)
-    /// to avoid float-overflow when intensities are large.</para>
+    /// <para>cpp uses <c>double</c> for the final <c>expIntSqSum*refIntSqSum</c> multiplication
+    /// explicitly (cpp DotProduct.cpp:95) to avoid float-overflow when intensities are large.
+    /// Inside the loop, however, the matched-MZ branch (cpp:68-70) multiplies <c>float*float</c>
+    /// and the unmatched branches use <c>pow</c> (which returns <c>double</c>) — we mirror that
+    /// asymmetry exactly to keep dot-product sums identical when used for tie-breaking the
+    /// representative-spectrum pick in <c>BlibFilter</c>.</para>
     /// </remarks>
     public static void Compare(Match match)
     {
@@ -56,9 +60,15 @@ public static class DotProduct
             var r = refp[j];
             if (e.Mz == r.Mz)
             {
-                expIntSqSum += (double)e.Intensity * e.Intensity;
-                refIntSqSum += (double)r.Intensity * r.Intensity;
-                expRefIntSum += (double)e.Intensity * r.Intensity;
+                // cpp DotProduct.cpp:68-70 — matched-MZ branch multiplies float*float (NOT
+                // pow's double), so the squared/product terms are truncated to float before
+                // accumulating into the double sum. Forcing this to double here was the
+                // source of a Filter-test tie-break divergence: TASEFDSAIAQDK picked id=116
+                // instead of cpp's id=50 because the average dot-product sums diverged in the
+                // 15th digit. Match cpp by computing float*float then promoting on add.
+                expIntSqSum += e.Intensity * e.Intensity;
+                refIntSqSum += r.Intensity * r.Intensity;
+                expRefIntSum += e.Intensity * r.Intensity;
                 matchedIons++;
                 i++;
                 j++;
